requirements.txt
# AI 프롬프트 생성기 - 필수 패키지
# Python 3.10 이상 필요

# Gemini API (새 버전)
google-genai>=1.57.0

# 이미지 처리
Pillow>=10.0.0

# 클립보드 복사
pyperclip>=1.8.0

# 환경 변수 관리
python-dotenv>=1.0.0
